feat(front): 보고서·사용자·워크스페이스 UX/권한 개편
- 관리자·클라이언트·PDF 보고서 뷰 Suspense 전환 + mount gate 로 로딩 race / hydration mismatch 해소
- SIR 주간 순위 비교 풀을 '동일 기간 리포트'로 수정 (rank/tier count 정합)
- SIR & 주가 차트 tick nice-snap 알고리즘 적용, 격자 ↔ 주가 레이블 정렬
- 보고서 섹션 탭 flex-1 균등 분할, client 뷰는 스크롤 시 상단 sticky + 섹션 전환 시 자동 스크롤 최상단
- UserDetailModal: 권한 스코프 강화 + 티어 변경 UI 추가 / StaffTable 담당 ws 컬럼 제거 / 계정 생성 super_admin 전용
- 워크스페이스 상세: 프로필 수정 / 블랙리스트 관리 버튼 분리, 디시 섹션 제거, 네이버 블로거(전역) 블랙리스트 지원
- /ops 역할 게이팅(super_admin + admin) + 사이드바 노출 필터
- Google Trends CSV 업로드 기능 (Route Handler + 클라이언트 파서, search_trends upsert)
- 워크스페이스 상세/목록 컴포넌트 분리 리팩터 (WeeklyReportCard, ReportListItem 등)
