+++ src/app/report-pdf/[workspaceId]/[reportId]/page.tsx
@@ -1,24 +1,35 @@
 'use client';
 
+import { Suspense, useEffect, useState } from 'react';
 import { useParams } from 'next/navigation';
-import { useIsFetching } from '@tanstack/react-query';
 import { Highlight } from '@/components/report/Highlight';
 import { OnlineReputation } from '@/components/report/OnlineReputation';
 import { TopContent } from '@/components/report/TopContent';
 import { RiskContent } from '@/components/report/RiskContent';
 import { Strategy } from '@/components/report/Strategy';
 import { Loading } from '@/components/ui/Loading';
-import { useReportInfo } from '@/hooks/report/useReportQuery';
+import { useReportInfoSuspense } from '@/hooks/report/useReportQuery';
 
 export default function ReportPdfPage() {
+  // 서버: QueryClient 캐시 비어있어 Loading fallback / 클라이언트: 기존 세션 캐시로 즉시 렌더
+  // → hydration mismatch 방지 위해 mount 전까지 서버와 동일하게 Loading 만 렌더
+  const [mounted, setMounted] = useState(false);
+  useEffect(() => setMounted(true), []); // eslint-disable-line react-hooks/set-state-in-effect
+  if (!mounted) return <Loading />;
+
+  return (
+    <Suspense fallback={<Loading />}>
+      <ReportPdfContent />
+    </Suspense>
+  );
+}
+
+function ReportPdfContent() {
   const params = useParams();
   const workspaceId = params?.workspaceId as string;
   const reportId = params?.reportId as string;
-  const isFetching = useIsFetching();
-  const { data: report } = useReportInfo(reportId);
+  const { data: report } = useReportInfoSuspense(reportId);
   const isDaily = report?.type === 'daily';
-
-  if (isFetching > 0) return <Loading />;
 
   return (
     <div className="p-8 bg-white min-h-screen">
