+++ src/components/chart/SirStockChart.tsx
@@ -36,6 +36,35 @@ function getWeekKey(fullDate: string): string {
   return monday.toISOString().slice(0, 10);
 }
 
+/** [min, max] 을 tickCount 개의 균등·"nice" 틱으로 스냅. 반환 배열 길이는 tickCount 고정. */
+function niceTicks(min: number, max: number, tickCount: number): number[] {
+  if (!(max > min)) return Array.from({ length: tickCount }, (_, i) => min + i);
+  const intervals = tickCount - 1;
+  const pickStep = (raw: number) => {
+    const mag = 10 ** Math.floor(Math.log10(raw));
+    const norm = raw / mag;
+    const factor = norm <= 1 ? 1 : norm <= 2 ? 2 : norm <= 2.5 ? 2.5 : norm <= 5 ? 5 : 10;
+    return factor * mag;
+  };
+  let step = pickStep((max - min) / intervals);
+  let lo = Math.floor(min / step) * step;
+  let hi = lo + step * intervals;
+  // hi < max 이면 step 을 다음 nice 값으로 승격해 전 구간 포함 보장
+  while (hi < max) {
+    step = pickStep(step * 1.01);
+    lo = Math.floor(min / step) * step;
+    hi = lo + step * intervals;
+  }
+  return Array.from({ length: tickCount }, (_, i) => lo + step * i);
+}
+
+/** 주가 틱 라벨 포맷 — step 크기에 따라 만 단위·소수점 자릿수 조정. */
+function formatPriceTick(v: number, step: number): string {
+  if (v < 10000) return v.toLocaleString();
+  const decimals = step >= 10000 ? 0 : step >= 1000 ? 1 : step >= 100 ? 2 : 3;
+  return `${(v / 10000).toFixed(decimals)}만`;
+}
+
 function aggregateWeekly(data: SirStockPoint[]) {
   const weeks = new Map<string, SirStockPoint[]>();
   for (const d of data) {
@@ -88,8 +117,16 @@ export function SirStockChart({
   const allPrices = chartInput.slice(-30)
     .flatMap((d) => [d.low_price, d.high_price])
     .filter((v): v is number => v != null);
-  const minPrice = allPrices.length ? Math.floor(Math.min(...allPrices) * 0.95) : 0;
-  const maxPrice = allPrices.length ? Math.ceil(Math.max(...allPrices) * 1.05) : 100;
+  const rawMin = allPrices.length ? Math.floor(Math.min(...allPrices) * 0.95) : 0;
+  const rawMax = allPrices.length ? Math.ceil(Math.max(...allPrices) * 1.05) : 100;
+
+  // 두 Y축을 동일 개수·동일 비율 위치에 고정 + nice tick 으로 주가 틱 라벨을 범위에 맞춰 스냅
+  const TICK_COUNT = 5;
+  const sirTicks = [0, 250, 500, 750, 1000];
+  const priceTicks = niceTicks(rawMin, rawMax, TICK_COUNT);
+  const minPrice = priceTicks[0];
+  const maxPrice = priceTicks[priceTicks.length - 1];
+  const priceStep = priceTicks[1] - priceTicks[0];
 
   return (
     <div className="outline-none **:outline-none">
@@ -115,6 +152,7 @@ export function SirStockChart({
               yAxisId="sir"
               orientation="left"
               domain={[0, 1000]}
+              ticks={sirTicks}
               tick={{ fontSize: 10, fill: 'var(--color-text-muted)' }}
               axisLine={{ stroke: '#d8dee9' }}
               tickLine={false}
@@ -132,11 +170,12 @@ export function SirStockChart({
               yAxisId="price"
               orientation="right"
               domain={[minPrice, maxPrice]}
+              ticks={priceTicks}
               tick={{ fontSize: 10, fill: 'var(--color-text-muted)' }}
               axisLine={{ stroke: '#d8dee9' }}
               tickLine={false}
               width={50}
-              tickFormatter={(v) => `${(v / 10000).toFixed(1)}만`}
+              tickFormatter={(v) => formatPriceTick(v, priceStep)}
               label={{
                 value: '주가',
                 angle: 90,
